fix static_cast<int>

+++ src/timbuk_parser-nobison.cc
@@ -1078,7 +1078,7 @@ Automata<Symbol> TimbukParser<Symbol>::from_line_to_automaton(std::string line) 
                 // }
             }
             for (const auto &t2 : aut2.transitions) {
-                int Q = aut.qubitNum + static_cast<int>(t2.first.symbol().qubit()); // we need to shift the qubit number
+                int Q = static_cast<int>(aut.qubitNum + static_cast<int>(t2.first.symbol().qubit())); // we need to shift the qubit number
                 if (t2.first.is_internal()) { // if the to-be-appended transition is internal, then
                     for (const auto &kv : t2.second) { // for each pair of top -> ...
                         auto top = kv.first;
